+++ OpenGL/src/game_objects/GameObject.h
@@ -19,7 +19,7 @@ public:
     GameObject(const std::string& name, int drawPriority, float x, float y);
     GameObject() = default;
     virtual ~GameObject() = default;
-    float CreationTime = glfwGetTime();
+    float CreationTime = (float)glfwGetTime();
     bool ShouldDestroy = false;
 
     virtual void setUpShader(Renderer& renderer);
